pyproject.toml
[tool.poetry]
name = "hello"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hello", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
